backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0
openai>=1.40.0
chromadb>=0.5.0
python-dotenv>=1.0.0
